US State Data from the NY Times

Update the state-level data from the New York Times on Covid-19

refresh_covid19nytimes_states()

Source

New York Times, "We’re Sharing Coronavirus Case Data for Every U.S. County" article and https://github.com/nytimes/covid-19-datadata repository

covid19R documentation

Value

A tibble object

date - The date in YYYY-MM-DD form

location - The name of the location as provided by the data source. The counties dataset provides county and state. They are combined and separated by a ,, and can be split by tidyr::separate(), if you wish.

location_type - The type of location using the covid19R controlled vocabulary. Nested locations are indicated by multiple location types being combined with a `_

location_standardized - A standardized location code using a national or international standard. In this case, FIPS state or county codes. See https://en.wikipedia.org/wiki/Federal_Information_Processing_Standard_state_code and https://en.wikipedia.org/wiki/FIPS_county_code for more

location_standardized_type The type of standardized location code being used according to the covid19R controlled vocabulary. Here we use fips_code

data_type - the type of data in that given row. Includes total_cases and total_deaths, cumulative measures of both.

value - number of cases of each data type

Details

Pulls from the open source data at https://www.nytimes.com/article/coronavirus-county-data-us.html including cumulative cases and deaths. Returns the data in the covid19R standard tidy format for easy use. See https://github.com/nytimes/covid-19-data for extensive methodology description and license.

Examples

if (FALSE) {

#update the data
covid19nytimes_states <- refresh_covid19nytimes_states()

}